Add files via upload
Project completed for University of Toronto SCS's Machine Learning Course.  As team lead I ensured our group remained on track to complete the project, I contributed significantly to the project, and I maintained a strong communication channel with our mentor Xuan Zhang.  The best model for our problem of creating clusters of Google fonts (of which there are over 1000 kinds) was to build and implement two consecutive neural networks, an autoencoder and a SOM.